Implementacion LOGUEO
Nuevo comando: LO
USO: LO -Fernando -1234
otras cosas que cambiaron:
Ahora cada hilo tiene asignado un objeto usuario, se modifico algunas variables y aspectos del codigo para que sea mas entendible.

+++ POO3-TP1-CHAT/src/Cliente_Servidor/ServidorHilo.java
@@ -8,51 +8,46 @@ import java.util.ArrayList;
 import java.util.logging.*;
 
 import domain.Conversaciones;
+import domain.Usuarios;
 
 
 public class ServidorHilo extends Thread {
 	
     private Socket socket;
-    private DataOutputStream dos;
-    private String accion;
-    public BufferedReader console;
-    private Servidor server;
-    private int threadID;
-    private ArrayList<Conversaciones> convsDelCliente;
+    private DataOutputStream escrituraConsCliente;
+    public BufferedReader lecturaConsCliente;
+    private Servidor server; //teniendo referencia al servidor, tendran acceso a la base de datos
+    private int threadID; //equivale al id del usuario
+    private Usuarios usuarioThread; //Objeto USUARIO que maneja el hilo
+    private ArrayList<Conversaciones> convsDelUsuario; //Conversaciones del usuario
     
     public ServidorHilo(Socket socket, int id) {
         this.socket = socket;
-        this.accion = "";
         try {
-            dos = new DataOutputStream(socket.getOutputStream());
-            this.console = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+            escrituraConsCliente = new DataOutputStream(socket.getOutputStream());
+            this.lecturaConsCliente = new BufferedReader(new InputStreamReader(socket.getInputStream()));
             this.server = Servidor.crear();
-            this.convsDelCliente = new ArrayList<Conversaciones>();
+            this.convsDelUsuario = new ArrayList<Conversaciones>();
+            this.usuarioThread = new Usuarios(); //creo el objeto usuario vacio
+            this.usuarioThread.setId_usuario_PK(0); //le asigno id= 0, la idea es que hasta que no se registre o loguee, esto actue como una restriccion.
+            //cuando se registra o loguea, se crean o recuperan los valores de la BDD y se asignan a este objeto. Si el user intenta conectarse a otro o crear
+            // o desconectarse, debe salirle un error.
         } catch (IOException ex) {
             Logger.getLogger(ServidorHilo.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
     
     
-    //Comando: CO
-    private void conectar() {
-    	try {
-			dos.writeUTF("#estableciendo conexion al server...\n");
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-    }
-    
     //Comando: UN -username
-    //Ahora este metodo es el responsable de asignar la ID al cliente
-    private void nombreUser() {
+    //comando de registro
+    private void nombreUser(String unBuffer) {
     	int idx = 0;
     	try {
-    		String aux = this.accion.substring(4, this.accion.length());
+    		String aux = unBuffer.substring(4, unBuffer.length());
     		idx = this.server.obtenerUsuarios().size() + 1000;
 			this.server.getConexionDB().consultaActualiza("INSERT INTO usuarios(id_usuario_PK, nombre_usuario, password_usuario) VALUES (" + idx + ", \'" + aux + "\', 1234);");
 			this.threadID = idx;
-			dos.writeUTF("#registro el username: " + aux + ", su ID es: " + idx + "\n");
+			escrituraConsCliente.writeUTF("#registro el username: " + aux + ", su ID es: " + idx + "\n");
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
@@ -61,23 +56,28 @@ public class ServidorHilo extends Thread {
     //Comando: CN -Id
     //Ahora el comando tiene la responsabilidad de crear un objeto conversacion y guardarlo en la lista local(arraylist<conversaciones>)
     //asi, cada hilo tiene el control de las conversaciones que ha iniciado
-    private void conectarCli() {
+    private int conectarCli(String unBuffer) {
     	int idx = 0;
     	try {
+    		if(this.usuarioThread.getId_usuario_PK() == 0) {
+        		escrituraConsCliente.writeUTF("No estas logueado, inicia sesion o registrate\n");
+        		return -1;
+        	}
     		idx = this.server.obtenerConversaciones().size() + 2000;
-    		String idDest = this.accion.substring(4, this.accion.length());
+    		String idDest = unBuffer.substring(4, unBuffer.length());
     		this.server.getConexionDB().consultaActualiza("INSERT INTO conversaciones(id_conversacion_pk, id_usuario1_fk, id_usuario2_fk) VALUES (" + idx + ", " + this.threadID + ",  " + idDest + ");");
-    		this.convsDelCliente.add(new Conversaciones(idx, this.threadID, Integer.parseInt(idDest)));
-			dos.writeUTF("#Has creado una nueva conversacion con el cliente ID: " + idDest + "\n");
+    		this.convsDelUsuario.add(new Conversaciones(idx, this.threadID, Integer.parseInt(idDest)));
+			escrituraConsCliente.writeUTF("#Has creado una nueva conversacion con el cliente ID: " + idDest + "\n");
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
+    	return 0;
     }
     
     //Comando: TX -mensaje
-    private void mensaje() {
+    private void mensaje(String unBuffer) {
     	try {
-			dos.writeUTF("#solicitud para enviar el mensaje: " + this.accion.substring(4, this.accion.length()) + "\n");
+			escrituraConsCliente.writeUTF("#solicitud para enviar el mensaje: " + unBuffer.substring(4, unBuffer.length()) + "\n");
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
@@ -92,41 +92,85 @@ public class ServidorHilo extends Thread {
     //donde el id_usuario2_fk (el user destino al que me conecte) coincida, y tambien restrinjo a que solo borre las conversaciones que haya iniciado YO
     //con ese user, asi evito borrar tambien aquellas conversaciones de otros usuarios que tenian tambien con el usuario al que me quiero desconectar.
     //Acto seguido recorro la lista de conversaciones local, hago la comprobacion y elimino los objetos conversaciones que coincidan
-    private void desconectarCli() {
+    private int desconectarCli(String unBuffer) {
     	try {
-    		int aux = Integer.parseInt(this.accion.substring(4, this.accion.length()));
-    		this.server.getConexionDB().consultaActualiza("DELETE FROM conversaciones WHERE Id_usuario2_fk = " + aux + "AND Id_usuario1_fk = " + this.threadID + ";");
-    		for(int i=0; i<this.convsDelCliente.size(); i++) {
-    			if(this.convsDelCliente.get(i).getId_usuario2_FK() == aux)
-    				this.convsDelCliente.remove(i);
+    		if(this.usuarioThread.getId_usuario_PK() == 0) {
+    			escrituraConsCliente.writeUTF("[DS] No estas logueado, inicia sesion o registrate\n");
+    			return -1;
     		}
-			dos.writeUTF("#te desconecto de la conversacion con el user ID: " + aux + "\n");
+    		if(this.convsDelUsuario.isEmpty()) {
+    			escrituraConsCliente.writeUTF("[DS] No tienes ninguna conversacion todavia\n");
+    			return -1;
+    		}
+    		int aux = Integer.parseInt(unBuffer.substring(4, unBuffer.length())); //parametro del comando
+    		boolean com = true;
+    		for(int i=0; i< this.convsDelUsuario.size(); i++) {
+    			if(this.convsDelUsuario.get(i).getId_usuario2_FK() == aux) {
+    				this.convsDelUsuario.remove(i);
+    				com = false;
+    			}
+    		}
+    		if(com) {
+    			escrituraConsCliente.writeUTF("[DS] No tienes ninguna conversacion con el ID que indicaste\n");
+    			return -1;
+    		}
+    		this.server.getConexionDB().consultaActualiza("DELETE FROM conversaciones WHERE Id_usuario2_fk = " + aux + "AND Id_usuario1_fk = " + this.threadID + ";");
+			escrituraConsCliente.writeUTF("#te desconecto de la conversacion con el user ID: " + aux + "\n");
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
+    	return 0;
     }
     
     //Metodo creado para debug, responde al comando QC
     private void consultarUsuarios() throws SQLException, IOException {
     	for(int i=0; i< this.server.obtenerUsuarios().size(); i++) {
-    		dos.writeUTF(this.server.obtenerUsuarios().get(i).getNombre_usuario() + " ");
-    		dos.writeUTF(this.server.obtenerUsuarios().get(i).getId_usuario_PK() + "\n");
+    		escrituraConsCliente.writeUTF(this.server.obtenerUsuarios().get(i).getNombre_usuario() + " ");
+    		escrituraConsCliente.writeUTF(this.server.obtenerUsuarios().get(i).getId_usuario_PK() + "\n");
     	}
     }
     
     //Otro metodo creado para debug, tambien responde al comando QC
     private void consultarConversaciones() throws SQLException, IOException {
     	for(int i=0; i<this.server.obtenerConversaciones().size(); i++) {
-    		dos.writeUTF(this.server.obtenerConversaciones().get(i).getId_conversacion_PK() + " ");
-    		dos.writeUTF(this.server.obtenerConversaciones().get(i).getId_usuario1_FK() + " ");
-    		dos.writeUTF(this.server.obtenerConversaciones().get(i).getId_usuario2_FK() + "\n");
+    		escrituraConsCliente.writeUTF(this.server.obtenerConversaciones().get(i).getId_conversacion_PK() + " ");
+    		escrituraConsCliente.writeUTF(this.server.obtenerConversaciones().get(i).getId_usuario1_FK() + " ");
+    		escrituraConsCliente.writeUTF(this.server.obtenerConversaciones().get(i).getId_usuario2_FK() + "\n");
     	}
+    }
+    
+    //busca en funcion del nombre y la contraseña
+    //modelo de comando:
+	//LO -Fernando -1234
+    private int loguear(String unBuffer) {
+    	try {
+    		String[] parts = unBuffer.split("-"); //divido la linea que recibo en partes
+    		String username = parts[1].replace(" ", ""); //asigno como username la segunda parte quitandole los espacios para evitar errores
+    		String passwd = parts[2].replace(" ", ""); //asigno como password la tercera parte quitandole los espacios para evitar errores
+    		//System.out.println("username: " + username.length() + ", pass: " + passwd.length());
+    		//recupero el usuario que cumple con el username y pass que recibi
+			ArrayList<Usuarios> usuarioLog = this.server.getConexionDB().recuperarUsuarios("SELECT * FROM usuarios WHERE nombre_usuario = \'" + username + "\' AND password_usuario = \'" + passwd + "\';");
+			//System.out.println(usuarioLog.size());
+			if(usuarioLog.isEmpty()) { //si lo que recibi esta vacio
+				escrituraConsCliente.writeUTF("[ERR_USER_NO_ENCONTRADO] No estas registrado en la base de datos o tu constrasenia no es correcta\n");
+				return -1;
+			}else if(usuarioLog.size() == 1){
+				this.usuarioThread = usuarioLog.get(0);
+				escrituraConsCliente.writeUTF("Bienvenido " + this.usuarioThread.getNombre_usuario() + "!!.\n");
+			}else {
+				escrituraConsCliente.writeUTF("error de ambiguedad en los datos\n");
+			}
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+    	return 0;
     }
     
     //Comando: EX
     private void desconectar(Socket unSoc) {
         try {
-        	dos.writeUTF("#finalizo la conexion\n");
+        	escrituraConsCliente.writeUTF("#finalizo la conexion\n");
         	System.out.println("Conexion saliente: "+unSoc);
             socket.close();
         } catch (IOException ex) {
@@ -140,55 +184,65 @@ public class ServidorHilo extends Thread {
     	boolean conectado = true;
         try {
         	//Mensaje de bienvenida
-        	dos.writeUTF("#Servidor escuchando...\n");
+        	escrituraConsCliente.writeUTF("#Servidor escuchando...\n");
         	
         	while(conectado) {
         		
         		//Leo la linea entera desde la consola de telnet
-        		this.accion = console.readLine();
+        		String lineaLeida = lecturaConsCliente.readLine();
        
-        		//Logica de resolucion de comandos:
-        		if(this.accion.equals("CO")){
-            	
-        			this.conectar();
-                  
-        		}else if(this.accion.substring(0,2).equals("UN")) {
-            	
-        			this.nombreUser();
-            	
-        		}else if(this.accion.substring(0,2).equals("CN")){
-            	
-        			this.conectarCli();
-            	
-        		}else if(this.accion.substring(0,2).equals("TX")) {
-            	
-        			this.mensaje();
-            	
-        		}else if(this.accion.substring(0,2).equals("DS")){
-            	
-        			this.desconectarCli();
-            	
-        		}else if(this.accion.equals("EX")) {
-            	
-        			this.desconectar(this.socket);
-        			conectado = false;
-            	
-        		}else if(this.accion.equals("QC")) {
+        		if(lineaLeida.length() < 2) { //si la longitud del comando es menor a dos, lanza error
+        			
+        			escrituraConsCliente.writeUTF("#[ERR_LONGITUD_MENOR_A_2] no entiendo el comando: " + lineaLeida + "\n");
+        			
+        		}else { //si no, se procede...
+        			
+        			//agarro los primeros dos caracteres de la linea
+        			String comando = lineaLeida.substring(0,2);
+        			
+        			//Logica de resolucion de comandos:
+            		if(comando.equals("LO")) {
+            			
+            			this.loguear(lineaLeida);
+            			
+            		} else if(comando.equals("UN")) {
                 	
-            		try {
-						//this.consultarUsuarios();
-						this.consultarConversaciones();
-						
-					} catch (SQLException e) {
-						e.printStackTrace();
-					}
+            			this.nombreUser(lineaLeida);
                 	
-            	}else {
-            	
-        			dos.writeUTF("#no entiendo el comando: " + this.accion + "\n");
-            	
+            		}else if(comando.equals("CN")){
+                	
+            			this.conectarCli(lineaLeida);
+                	
+            		}else if(comando.equals("TX")) {
+                	
+            			this.mensaje(lineaLeida);
+                	
+            		}else if(comando.equals("DS")){
+                	
+            			this.desconectarCli(lineaLeida);
+                	
+            		}else if(comando.equals("EX")) {
+                	
+            			this.desconectar(this.socket);
+            			conectado = false;
+                	
+            		}else if(comando.equals("QC")) {
+                    	
+                		try {
+    						this.consultarUsuarios();
+    						this.consultarConversaciones();
+    						
+    					} catch (SQLException e) {
+    						e.printStackTrace();
+    					}
+                    	
+                	}else {
+                	
+            			escrituraConsCliente.writeUTF("#[ERR_COMANDO_NO_EXISTE] no entiendo el comando: " + lineaLeida + "\n");
+                	
+            		}
         		}
-            
+        		
         	}
             
         } catch (IOException ex) {
